Refactor by component.
